Split search term by type, e.g. keyword, target
