+++ src/api/amazon-ads/index.js
@@ -1,3 +1,4 @@
 export * from './reports'
 export * from './keywords'
 export * from './search-terms'
+export * from './product-targets'
