+++ src/api/amazon-ads/product-targets.js
@@ -1,0 +1,7 @@
+import { call } from '../core'
+
+export async function getProductTarget(asin, adGroupId) {
+  return call(`amazon/ads/product-targets/${asin}?adGroupId=${adGroupId}`, {
+    method: 'GET'
+  })
+}
