--- src/components/Loader.jsx
@@ -1,12 +1,0 @@
-import React from 'react'
-import { Loader as MantineLoader } from '@mantine/core'
-
-function Loader() {
-  return (
-    <div>
-      <MantineLoader size="sm" />
-    </div>
-  )
-}
-
-export default Loader
